+++ test/controllers/api/v1/subjects_controller_test.rb
@@ -25,7 +25,7 @@ class Api::V1::SubjectsControllerTest < ActionDispatch::IntegrationTest
 
   test "should create subject" do
     assert_difference("Subject.count") do
-      post api_v1_subjects_path(
+      post api_v1_create_subject_path(
         authentication_token: @project.id_and_token,
         subject: {
           subject_code: "S00001",
@@ -52,6 +52,22 @@ class Api::V1::SubjectsControllerTest < ActionDispatch::IntegrationTest
     assert_not_nil assigns(:subject)
     assert_not_nil assigns(:event)
     assert_not_nil assigns(:subject_event)
+    assert_response :success
+  end
+
+  test "should create subject sheet" do
+    assert_difference("Sheet.count") do
+      post api_v1_create_sheet_path(
+        authentication_token: @project.id_and_token,
+        id: @subject,
+        subject_event_id: subject_events(:api_one_api_event),
+        design_id: designs(:api_design).id,
+        format: "json"
+      )
+    end
+    assert_not_nil assigns(:project)
+    assert_not_nil assigns(:subject)
+    assert_not_nil assigns(:sheet)
     assert_response :success
   end
 end