+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 - **API Changes**
   - Added API to generate and retrieve a project's subjects and events
 
+### Bug Fix
+- Subjects with an event but no sheets now correctly display the event
+
 ## 0.54.1 (July 13, 2017)
 
 ### Bug Fix